+++ dtslint/ts3.5/Encoder.ts
@@ -9,7 +9,7 @@ const BooleanToNumber: E.Encoder<number, boolean> = {
   encode: (b) => (b ? 1 : 0)
 }
 
-export const OfTest = E.type({ a: E.id<string>(), b: E.type({ c: NumberToString }) })
+export const OfTest = E.struct({ a: E.id<string>(), b: E.struct({ c: NumberToString }) })
 
 //
 // TypeOf
@@ -27,9 +27,9 @@ export type OfTestOutput = E.OutputOf<typeof OfTest> // $ExpectType { a: string;
 E.nullable(NumberToString) // $ExpectType Encoder<string | null, number | null>
 
 //
-// type
+// struct
 //
-E.type({ a: E.type({ b: NumberToString }) }) // $ExpectType Encoder<{ a: { b: string; }; }, { a: { b: number; }; }>
+E.struct({ a: E.struct({ b: NumberToString }) }) // $ExpectType Encoder<{ a: { b: string; }; }, { a: { b: number; }; }>
 
 //
 // partial
@@ -56,19 +56,19 @@ E.tuple(NumberToString, BooleanToNumber) // $ExpectType Encoder<[string, number]
 //
 // intersection
 //
-pipe(E.type({ a: NumberToString }), E.intersect(E.type({ b: BooleanToNumber }))) // $ExpectType Encoder<{ a: string; } & { b: number; }, { a: number; } & { b: boolean; }>
+pipe(E.struct({ a: NumberToString }), E.intersect(E.struct({ b: BooleanToNumber }))) // $ExpectType Encoder<{ a: string; } & { b: number; }, { a: number; } & { b: boolean; }>
 
 //
 // sum
 //
-const S1 = E.type({ _tag: E.id<'A'>(), a: NumberToString })
-const S2 = E.type({ _tag: E.id<'B'>(), b: BooleanToNumber })
+const S1 = E.struct({ _tag: E.id<'A'>(), a: NumberToString })
+const S2 = E.struct({ _tag: E.id<'B'>(), b: BooleanToNumber })
 const sum = E.sum('_tag')
 
 // $ExpectType Encoder<{ _tag: "A"; a: string; } | { _tag: "B"; b: number; }, { _tag: "A"; a: number; } | { _tag: "B"; b: boolean; }>
 sum({ A: S1, B: S2 })
 
-const S3 = E.type({ _tag: E.id<'C'>(), c: E.id<string>() })
+const S3 = E.struct({ _tag: E.id<'C'>(), c: E.id<string>() })
 
 //
 // lazy
@@ -90,14 +90,14 @@ interface BOut {
   as: Array<AOut>
 }
 const A: E.Encoder<AOut, A> = E.lazy(() =>
-  E.type({
+  E.struct({
     a: NumberToString,
     bs: E.array(B)
   })
 )
 
 const B: E.Encoder<BOut, B> = E.lazy(() =>
-  E.type({
+  E.struct({
     b: BooleanToNumber,
     as: E.array(A)
   })