+++ test/TaskDecoder.ts
@@ -137,7 +137,7 @@ describe('UnknownTaskDecoder', () => {
 
   it('withMessage', async () => {
     const decoder = pipe(
-      _.type({
+      _.struct({
         name: _.string,
         age: _.number
       }),
@@ -196,30 +196,30 @@ describe('UnknownTaskDecoder', () => {
     })
   })
 
-  describe('type', () => {
+  describe('struct', () => {
     it('should decode a valid input', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: _.string
       })
       assert.deepStrictEqual(await decoder.decode({ a: 'a' })(), D.success({ a: 'a' }))
     })
 
     it('should strip additional fields', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: _.string
       })
       assert.deepStrictEqual(await decoder.decode({ a: 'a', b: 1 })(), D.success({ a: 'a' }))
     })
 
     it('should not strip fields corresponding to undefined values', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: undef
       })
       assert.deepStrictEqual(await decoder.decode({})(), D.success({ a: undefined }))
     })
 
     it('should reject an invalid input', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: _.string
       })
       assert.deepStrictEqual(await decoder.decode(undefined)(), D.failure(undefined, 'Record<string, unknown>'))
@@ -230,7 +230,7 @@ describe('UnknownTaskDecoder', () => {
     })
 
     it('should collect all errors', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: _.string,
         b: _.number
       })
@@ -254,7 +254,7 @@ describe('UnknownTaskDecoder', () => {
           return 'b'
         }
       }
-      const decoder = _.type({ a: _.string, b: _.string })
+      const decoder = _.struct({ a: _.string, b: _.string })
       assert.deepStrictEqual(await decoder.decode(new A())(), D.success({ a: 'a', b: 'b' }))
     })
   })
@@ -465,7 +465,7 @@ describe('UnknownTaskDecoder', () => {
 
   describe('intersect', () => {
     it('should decode a valid input', async () => {
-      const decoder = pipe(_.type({ a: _.string }), _.intersect(_.type({ b: _.number })))
+      const decoder = pipe(_.struct({ a: _.string }), _.intersect(_.struct({ b: _.number })))
       assert.deepStrictEqual(await decoder.decode({ a: 'a', b: 1 })(), D.success({ a: 'a', b: 1 }))
     })
 
@@ -475,7 +475,7 @@ describe('UnknownTaskDecoder', () => {
     })
 
     it('should accumulate all errors', async () => {
-      const decoder = pipe(_.type({ a: _.string }), _.intersect(_.type({ b: _.number })))
+      const decoder = pipe(_.struct({ a: _.string }), _.intersect(_.struct({ b: _.number })))
       assert.deepStrictEqual(
         await decoder.decode({ a: 'a' })(),
         E.left(FS.of(DE.key('b', DE.required, FS.of(DE.leaf(undefined, 'number')))))
@@ -500,16 +500,16 @@ describe('UnknownTaskDecoder', () => {
     const sum = _.sum('_tag')
 
     it('should decode a valid input', async () => {
-      const A = _.type({ _tag: _.literal('A'), a: _.string })
-      const B = _.type({ _tag: _.literal('B'), b: _.number })
+      const A = _.struct({ _tag: _.literal('A'), a: _.string })
+      const B = _.struct({ _tag: _.literal('B'), b: _.number })
       const decoder = sum({ A, B })
       assert.deepStrictEqual(await decoder.decode({ _tag: 'A', a: 'a' })(), D.success({ _tag: 'A', a: 'a' }))
       assert.deepStrictEqual(await decoder.decode({ _tag: 'B', b: 1 })(), D.success({ _tag: 'B', b: 1 }))
     })
 
     it('should reject an invalid input', async () => {
-      const A = _.type({ _tag: _.literal('A'), a: _.string })
-      const B = _.type({ _tag: _.literal('B'), b: _.number })
+      const A = _.struct({ _tag: _.literal('A'), a: _.string })
+      const B = _.struct({ _tag: _.literal('B'), b: _.number })
       const decoder = sum({ A, B })
       assert.deepStrictEqual(await decoder.decode(null)(), D.failure(null, 'Record<string, unknown>'))
       assert.deepStrictEqual(
@@ -537,7 +537,7 @@ describe('UnknownTaskDecoder', () => {
   }
 
   const lazyDecoder: _.TaskDecoder<unknown, A> = _.lazy('A', () =>
-    pipe(_.type({ a: NumberFromString }), _.intersect(_.partial({ b: lazyDecoder })))
+    pipe(_.struct({ a: NumberFromString }), _.intersect(_.partial({ b: lazyDecoder })))
   )
 
   describe('lazy', () => {
@@ -581,7 +581,7 @@ describe('UnknownTaskDecoder', () => {
 
   describe('draw', () => {
     it('draw', async () => {
-      const decoder = _.type({
+      const decoder = _.struct({
         a: _.string,
         b: _.number,
         c: _.array(_.boolean),